compliance/requirements.txt
PyMuPDF
streamlit
PyPDF2
python-docx
Pillow
pytesseract
fitz
fpdf
groq
json
os